refactoring, styles
