+++ docs/src/prototypes/chatMessages/ControlMessages/ControlMessage.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react'
-import { Ref, ChatMessageProps, Chat } from '@stardust-ui/react'
+import { Ref, ChatMessageProps, Chat, Icon } from '@stardust-ui/react'
 
 interface ControlMessageProps {
   focused?: boolean
+  icon?: boolean
   message: ChatMessageProps
 }
 class ControlMessage extends React.Component<ControlMessageProps> {
@@ -16,12 +17,20 @@ class ControlMessage extends React.Component<ControlMessageProps> {
 
   render() {
     return (
-      <Ref innerRef={this.messageRef}>
-        <Chat.Message
-          {...this.props.message}
-          styles={{ padding: 0, marginLeft: '10px', backgroundColor: '#f3f2f1', fontSize: '14px' }}
-        />
-      </Ref>
+      <>
+        {this.props.icon ? <Icon name="participant-add" /> : null}
+        <Ref innerRef={this.messageRef}>
+          <Chat.Message
+            {...this.props.message}
+            styles={{
+              padding: 0,
+              marginLeft: '10px',
+              backgroundColor: '#f3f2f1',
+              fontSize: '14px',
+            }}
+          />
+        </Ref>
+      </>
     )
   }
 }
