+++ docs/src/prototypes/chatMessages/ControlMessages/GroupControlMessages.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react'
 import * as _ from 'lodash'
+import * as keyboardKey from 'keyboard-key'
 
-import { List, ChatMessageProps, chatBehavior, Accessibility } from '@stardust-ui/react'
+import { List, chatBehavior, Accessibility, Flex, Icon } from '@stardust-ui/react'
 import ControlMessage from './ControlMessage'
 
 const overridenChatBehavior: Accessibility<any> = props => {
@@ -24,27 +25,79 @@ const overridenChatBehavior: Accessibility<any> = props => {
   return behaviorData
 }
 
-interface GroupControlMessagesProps {
-  items: ChatMessageProps[]
-}
+const GroupControlMessages = () => {
+  const [expanded, setExpanded] = React.useState(false)
+  const [focused, setFocused] = React.useState(false)
 
-class GroupControlMessages extends React.Component<GroupControlMessagesProps> {
-  renderItems = () => {
-    return _.map(this.props.items, item => {
+  const renderItems = () => {
+    return _.map(groupControlMessageItems, item => {
       return {
         content: <ControlMessage message={item} />,
-        styles: {
-          padding: 0,
-          display: 'block',
-          minHeight: '25px',
-        },
+        styles: { padding: 0, display: 'block', minHeight: '25px' },
       }
     })
   }
 
-  render() {
-    return <List accessibility={overridenChatBehavior} items={this.renderItems()} />
-  }
+  return (
+    <Flex
+      onKeyDown={e => {
+        const eventCode = keyboardKey.getCode(e)
+        if (eventCode === keyboardKey.Enter) {
+          setExpanded(true)
+        }
+        if (eventCode === keyboardKey.Escape) {
+          setExpanded(false)
+          setFocused(true)
+        }
+      }}
+    >
+      <Icon
+        name={expanded ? 'stardust-arrow-down' : 'stardust-arrow-end'}
+        onClick={() => setExpanded(!expanded)}
+      />
+      <Icon name="participant-add" />
+      {expanded ? (
+        <List accessibility={overridenChatBehavior} items={renderItems()} />
+      ) : (
+        <ControlMessage focused={focused} message={controlMessage} />
+      )}
+    </Flex>
+  )
 }
 
 export default GroupControlMessages
+
+const groupControlMessageItems = [
+  {
+    key: 'joe-doe1',
+    content: (
+      <div>
+        <a href="/">John Doe</a> has added <a href="/">Jane Doe1</a> to the team
+      </div>
+    ),
+  },
+  {
+    key: 'joe-doe2',
+    content: (
+      <div>
+        <a href="/">John Doe</a> has added <a href="/">Jane Doe2</a> to the team
+      </div>
+    ),
+  },
+  {
+    key: 'joe-doe3',
+    content: (
+      <div>
+        <a href="/">John Doe</a> has added <a href="/">Jane Doe3</a> to the team
+      </div>
+    ),
+  },
+]
+
+const controlMessage = {
+  content: (
+    <div>
+      <a href="/">John Doe</a> has added <a href="/">Jane Doe1</a> and 2 other to the team
+    </div>
+  ),
+}
